Refactored to command+argument structure
This is an attempt to introduce more modularity. Currenty
the codebase is structured around command model where
a new class is introduced for each new operation. I suggest
we change this to something better matching vim model.

Many vim commands take the following form: <operator><text_object | motion>.
For example the following commands:
	caw, daw, yaw, vaw, g~aw
all perform their actions on the same text object: a word under caret.

This commit adds a single command ChangeCommand which takes
as argument a function which in turn returns a CommandRange.
The ChangeCommand acts on this range, i.e. cuts the text and
switches the text editor to Insert mode.

I made the argument a function to underline its purpose of
being a simple calculator. It should be a pure function in
mathematical sense: doesn't change any state, doesn't have
any side effects.

Only "a block" (a(, a[, a{) and "quoted string" text objects are
implemented in this commit. No refactoring to existing commands
was performed, but this should be trivial with the new architecture.